requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.22.0
azure-storage-queue==12.10.0
python-dotenv==1.0.1
requests==2.32.3
